int calculate

+++ src/BmiService.java
@@ -1,6 +1,6 @@
 public class BmiService {
 
-    public double calculate(double height, double weight) {
+    public int calculate(double height, double weight) {
         return (int) (weight / (height * height));
     }
 }
